feat(report): Initial work on printable report. Includes major refactoring of Site component
