+++ client/components/site.js
@@ -1,14 +1,9 @@
-import hslToHex from 'colorvert/hsl/hex'
-import {toCoordinates, toLeaflet} from '@conveyal/lonlat'
+import {toCoordinates} from '@conveyal/lonlat'
 import humanizeDuration from 'humanize-duration'
-import {geom, io, precision, simplify} from 'jsts'
-import {Browser, icon, latLngBounds, point} from 'leaflet'
 import React, {Component, PropTypes} from 'react'
 import {Button, ButtonGroup, Col, ControlLabel, FormGroup, Grid, Panel,
   ProgressBar, Row, Tab, Table, Tabs} from 'react-bootstrap'
 import {BootstrapTable, TableHeaderColumn} from 'react-bootstrap-table'
-import {Circle, GeoJSON, Map as LeafletMap, Marker, Popup, TileLayer} from 'react-leaflet'
-import Heatmap from 'react-leaflet-heatmap-layer'
 import Combobox from 'react-widgets/lib/Combobox'
 import Slider from 'rc-slider'
 import distance from '@turf/distance'
@@ -17,14 +12,13 @@ import BackButton from '../containers/back-button'
 import ButtonLink from './button-link'
 import FieldGroup from './fieldgroup'
 import Icon from './icon'
-import Legend from './legend'
-import MarkerCluster from './marker-cluster'
-import {actUponConfirmation, arrayCountRenderer, humanizeDistance} from '../utils'
+import {actUponConfirmation, arrayCountRenderer, humanizeDistance,
+  formatDistance, formatPercent, formatPercentAsStr} from '../utils'
 import messages from '../utils/messages'
-import {modeshareStats} from '../utils/settings'
+import SiteMap from './site-map'
+import SiteInfographic from './site-infographic'
 
-const geoJsonReader = new io.GeoJSONReader()
-const geoJsonWriter = new io.GeoJSONWriter()
+import { processSite } from './site-common'
 
 export default class Site extends Component {
   static propTypes = {
@@ -55,7 +49,7 @@ export default class Site extends Component {
       commuterRingRadius: 1,
       isochroneCutoff: 7200,
       rideMatchMapStyle: 'normal',
-      mapState: 'STANDARD' // STANDARD / FULLSCREEN / HIDDEN
+      mapDisplayMode: 'STANDARD' // STANDARD / FULLSCREEN / HIDDEN
     }
     this._loadDataIfNeeded(this.props)
   }
@@ -142,10 +136,10 @@ export default class Site extends Component {
     this.setState({ activeTab: selectedTab })
   }
 
-  _setMapState (mapState) {
-    this.setState({ mapState })
+  _setMapDisplayMode (mapDisplayMode) {
+    this.setState({ mapDisplayMode })
     setTimeout(() => {
-      this.refs['leafletMap'].leafletElement.invalidateSize(false)
+      this.refs['map'].resized()
     }, 100)
   }
 
@@ -243,76 +237,6 @@ export default class Site extends Component {
     }
   }
 
-  _mapSitesAndCommuters = () => {
-    const {commuters, isMultiSite, site, sites} = this.props
-    const {selectedCommuter} = this.state
-    const commuterMarkers = []
-    const siteMarkers = []
-    let sitesToMakeMarkersFor
-
-    if (isMultiSite) {
-      sitesToMakeMarkersFor = sites
-    } else {
-      sitesToMakeMarkersFor = [site]
-    }
-
-    const firstSiteCoordinates = toLeaflet(sitesToMakeMarkersFor[0].coordinate)
-    const bounds = latLngBounds([firstSiteCoordinates, firstSiteCoordinates])
-
-    sitesToMakeMarkersFor.forEach((siteToMakeMarkerFor) => {
-      const sitePosition = toLeaflet(siteToMakeMarkerFor.coordinate)
-
-      // add site marker
-      siteMarkers.push(
-        <Marker
-          key={`site-marker-${siteToMakeMarkerFor._id}`}
-          position={sitePosition}
-          zIndexOffset={9000}
-        />
-      )
-
-      bounds.extend(sitePosition)
-    })
-
-    // add all commuters to site
-    commuters.forEach((commuter) => {
-      if (commuter.coordinate.lat === 0) return  // don't include commuters not geocoded yet
-      const commuterPosition = toLeaflet(commuter.coordinate)
-      const isSelectedCommuter = selectedCommuter && selectedCommuter._id === commuter._id
-      commuterMarkers.push(
-        <Marker
-          icon={isSelectedCommuter ? homeIconSelected : homeIcon}
-          key={`commuter-marker-${commuter._id}`}
-          onClick={() => this._handleSelectCommuter(commuter, true)}
-          position={commuterPosition}
-          zIndexOffset={1234}
-          >
-          <Popup
-            offset={homeIconSelectedOffset}
-            >
-            <h4>{commuter.name}</h4>
-          </Popup>
-        </Marker>
-      )
-      bounds.extend(commuterPosition)
-    })
-
-    // return only site marker if no commuters or commuters haven't loaded yet
-    if (commuterMarkers.length === 0 && siteMarkers.length === 1) {
-      return {
-        siteMarkers,
-        position: firstSiteCoordinates,
-        zoom: 11
-      }
-    }
-
-    return {
-      bounds,
-      commuterMarkers,
-      siteMarkers
-    }
-  }
-
   _onDeleteCommuterClick (commuter) {
     const doDelete = () => this.props.deleteCommuter(commuter)
     actUponConfirmation(messages.commuter.deleteConfirmation, doDelete)
@@ -327,10 +251,15 @@ export default class Site extends Component {
       commuterRingRadius,
       isochroneCutoff,
       rideMatchMapStyle,
-      selectedCommuter
+      selectedCommuter,
+      mapDisplayMode
     } = this.state
 
     const hasCommuters = commuters.length > 0
+
+    const processed = processSite(commuters, analysisMode)
+
+    /*const hasCommuters = commuters.length > 0
     const pctGeocoded = formatPercent(commuters.reduce((accumulator, commuter) => {
       return accumulator + (commuter.geocodeConfidence !== -1 ? 1 : 0)
     }, 0) / commuters.length)
@@ -338,147 +267,13 @@ export default class Site extends Component {
       return accumulator + (commuter.modeStats ? 1 : 0)
     }, 0) / commuters.length)
     const allCommutersGeocoded = pctGeocoded === 100
-    const allCommutersStatsCalculated = pctStatsCalculated === 100
-
-    /************************************************************************
-     map stuff
-    ************************************************************************/
-    const mapLegendProps = {
-      html: '<h4>Legend</h4><table><tbody>',
-      position: 'bottomright'
-    }
-
-    // add marker to legend
-    const siteIconUrl = 'https://unpkg.com/leaflet@1.0.2/dist/images/marker-icon-2x.png'
-    mapLegendProps.html += `<tr><td><img src="${siteIconUrl}" style="width: 25px;"/></td><td>Site</td></tr>`
-
-    const {bounds, commuterMarkers, position, siteMarkers, zoom} = this._mapSitesAndCommuters()
-    const clusterMarkers = []
-    const commuterRings = []
-
-    if (hasCommuters) {
-      if (activeTab === 'ridematches') {
-        if (rideMatchMapStyle === 'normal') {
-          mapLegendProps.html += `<tr><td><img src="${homeIconUrl}" /></td><td>Commuter</td></tr>`
-        } else if (rideMatchMapStyle === 'marker-clusters') {
-          mapLegendProps.html += `<tr>
-            <td>
-              <img src="${homeIconUrl}" />
-            </td>
-            <td>Single Commuter</td>
-          </tr>
-          <tr>
-            <td>
-              <img src="${process.env.STATIC_HOST}assets/cluster.png" style="width: 40px;"/>
-            </td>
-            <td>Cluster of Commuters</td>
-          </tr>`
-
-          commuterMarkers.forEach((marker) => {
-            clusterMarkers.push({
-              id: marker.key,
-              latLng: marker.props.position,
-              markerOptions: marker.props,
-              onClick: marker.props.onClick
-            })
-          })
-        } else if (rideMatchMapStyle === 'heatmap') {
-          mapLegendProps.html += `<tr>
-            <td
-              rowspan="2"
-              style="background-image: url(${process.env.STATIC_HOST}assets/heatmap-gradient.png);
-                background-size: contain;"
-              />
-            <td>Less Commuters</td>
-          </tr>
-          <tr>
-            <td>More Commuters</td>
-          </tr>`
-        } else if (rideMatchMapStyle === 'commuter-rings') {
-          mapLegendProps.html += `<tr>
-            <td>
-              <img src="${homeIconUrl}" />
-            </td>
-            <td>Commuter</td>
-          </tr>
-          <tr>
-            <td>
-              <div style="border: 3px solid #3388ff; border-radius: 20px; overflow: hidden;">
-                <div style="background-color: #3388ff; opacity: 0.2; height: 25px;">
-                  &nbsp;
-                </div>
-              </div>
-            </td>
-            <td>${humanizeDistance(commuterRingRadius, 2)} Radius</td>
-          </tr>`
-
-          const meters = commuterRingRadius * 1609.34
-
-          commuterMarkers.forEach((marker) => {
-            commuterRings.push(
-              <Circle
-                center={marker.props.position}
-                key={`commuter-circle-${marker.key}`}
-                radius={meters}
-                />
-            )
-          })
-        }
-      } else {
-        mapLegendProps.html += `<tr><td><img src="${homeIconUrl}" /></td><td>Commuter</td></tr>`
-      }
-    }
-
-    if ((!(activeTab === 'ridematches') || (
-      activeTab === 'ridematches' && rideMatchMapStyle !== 'heatmap'
-    )) && selectedCommuter) {
-      mapLegendProps.html += `<tr>
-        <td>
-          <img src="${homeIconSelectedUrl}" />
-        </td>
-        <td>${selectedCommuter.name}</td>
-      </tr>`
-    }
-
-    // isochrones
-    const isochrones = []
-    if (!isMultiSite &&
-      activeTab === 'analysis' &&
-      site.calculationStatus === 'successfully') {
-      // travel times calculated successfully
-      const curIsochrones = getIsochrones({
-        analysisMapStyle,
-        analysisMode,
-        isochroneCutoff,
-        polygonStore,
-        site
-      })
-      curIsochrones
-        .filter((isochrone) => isochrone.properties.time <= isochroneCutoff)
-        .forEach((isochrone) => {
-          const geojsonProps = {
-            data: Object.assign(isochrone, { type: 'Feature' }),
-            key: `isochrone-${analysisMapStyle}-${analysisMode}-${isochrone.properties.time}`,
-            onEachFeature
-          }
-
-          if (isochroneStyleStrategies[analysisMapStyle]) {
-            Object.assign(geojsonProps, isochroneStyleStrategies[analysisMapStyle])
-          }
-
-          isochrones.push(
-            <GeoJSON {...geojsonProps} />
-          )
-        })
-
-      mapLegendProps.html += getIsochroneLegendHtml({ analysisMapStyle, isochroneCutoff, analysisMode })
-    }
+    const allCommutersStatsCalculated = pctStatsCalculated === 100*/
 
     /************************************************************************
      summary tab stuff
     ************************************************************************/
 
-    const summaryStats = {}
+    /*const summaryStats = {}
 
     if (allCommutersStatsCalculated) {
       let numWith60MinTransit = 0
@@ -498,11 +293,7 @@ export default class Site extends Component {
 
       summaryStats.pctWith60MinTransit = formatPercentAsStr(numWith60MinTransit / commuters.length)
       summaryStats.pctWith30MinBike = formatPercentAsStr(numWith30MinBike / commuters.length)
-    }
-
-    const regionalTransitModePct = formatPercentAsStr(modeshareStats.transit)
-    const regionalOtherModePct = formatPercentAsStr(modeshareStats.other)
-    const regionalCarpoolModePct = formatPercentAsStr(modeshareStats.carpool)
+    }*/
 
     /************************************************************************
      commuter tab stuff
@@ -530,7 +321,12 @@ export default class Site extends Component {
     /************************************************************************
      analysis tab stuff
     ************************************************************************/
-    const analysisModeStatsLookup = {}
+
+    const analysisSliderStepAndMin = (
+      getIsochroneStrategies[analysisMapStyle] === '15-minute isochrones'
+    ) ? 900 : 300
+
+    /*const analysisModeStatsLookup = {}
     const analysisSliderStepAndMin = (
       getIsochroneStrategies[analysisMapStyle] === '15-minute isochrones'
     ) ? 900 : 300
@@ -567,13 +363,13 @@ export default class Site extends Component {
           cumulative,
           cumulativePct: cumulative / commuters.length
         }
-      })
+      })*/
 
     /************************************************************************
      ridematches tab stuff
     ************************************************************************/
     // only do this if all commuters are geocoded
-    const ridematches = {}
+    /*const ridematches = {}
     const ridematchingAggregateTable = []
 
     const addRidematch = (commuterA, commuterB, distance) => {
@@ -648,9 +444,7 @@ export default class Site extends Component {
 
       const upToOneMileBinIdx = 2
       summaryStats.pctWithRidematch = formatPercentAsStr(ridematchingAggregateTable[upToOneMileBinIdx].cumulativePct)
-    }
-
-    mapLegendProps.html += '</tbody></table>'
+    }*/
 
     return (
       <Grid>
@@ -692,7 +486,7 @@ export default class Site extends Component {
             Content
           ***************************/}
           {!hasCommuters &&
-            <Col xs={this.state.mapState === 'HIDDEN' ? 12 : 7}>
+            <Col xs={this.state.mapDisplayMode === 'HIDDEN' ? 12 : 7}>
               {isMultiSite &&
                 <p>None of the sites in this Multi-Site Analysis have any commuters!  Add commuters to specific sites.</p>
               }
@@ -705,15 +499,15 @@ export default class Site extends Component {
             </Col>
           }
           {hasCommuters &&
-            <Col xs={this.state.mapState === 'HIDDEN' ? 12 : 7}>
+            <Col xs={this.state.mapDisplayMode === 'HIDDEN' ? 12 : 7}>
             <Tabs
               activeKey={activeTab}
               id='site-tabs'
               onSelect={this._handleTabSelect}
               >
-              {this.state.mapState === 'HIDDEN' &&
+              {this.state.mapDisplayMode === 'HIDDEN' &&
                 <div style={{ position: 'absolute', right: '15px', top: '0px' }}>
-                  <Button bsSize='small' onClick={() => this.setState({ mapState: 'STANDARD' })}>
+                  <Button bsSize='small' onClick={() => this.setState({ mapDisplayMode: 'STANDARD' })}>
                     <Icon type='map' /> Show Map
                   </Button>
                 </div>
@@ -722,123 +516,45 @@ export default class Site extends Component {
                 {/***************************
                   Summary Tab
                 ***************************/}
-                {!allCommutersGeocoded &&
+                {!processed.allCommutersGeocoded &&
                   <ProgressBar
                     striped
-                    now={pctGeocoded}
+                    now={processed.pctGeocoded}
                     label='Geocoding Commuters'
                     />
                 }
-                {allCommutersGeocoded && !allCommutersStatsCalculated &&
+                {processed.allCommutersGeocoded && !processed.allCommutersStatsCalculated &&
                   <ProgressBar
                     striped
-                    now={pctStatsCalculated}
+                    now={processed.pctStatsCalculated}
                     label='Analyzing Commutes'
                     />
                 }
-                {allCommutersGeocoded && allCommutersStatsCalculated &&
-                  <Row>
-                    <Row>
-                      <Col xs={6} className='infographic-site-column'>
-                        <h4>Total Commuters</h4>
-                        <div className='infographic-well' style={{backgroundColor: '#51992e'}}>
-                          <Icon type='group' />
-                          <span className='number'>{commuters.length}</span>
-                        </div>
-                        <p>
-                          {commuters.length}
-                          {isMultiSite
-                            ? ' commuters are at these sites.'
-                            : ' commuters are at this site.'}</p>
-                      </Col>
-                      <Col xs={6} className='infographic-site-column'>
-                        <h4>Transit Commute</h4>
-                        <div
-                          className='infographic-well'
-                          style={infographicBackground('#3b90c6', summaryStats.pctWith60MinTransit)}
-                          >
-                          <Icon type='bus' />
-                          <span className='number-pct'>{summaryStats.pctWith60MinTransit}</span>
-                        </div>
-                        <p>
-                          {summaryStats.pctWith60MinTransit} of commuters at
-                          {isMultiSite ? ' these sites ' : ' this site '}
-                          are within a 60 minute transit ride.
-                        </p>
-                      </Col>
-                    </Row>
+                {processed.allCommutersGeocoded && processed.allCommutersStatsCalculated &&
 
+                  <Row className='summary-tab'>
                     <Row>
-                      <Col xs={6} className='infographic-modeshare-container'>
-                        <h4>Compare Against Washington, DC Averages</h4>
+                      <Col xs={12}>
+                        <Panel header={(<h3>Printable Report</h3>)}>
+                          <Col xs={5}>
+                            <ButtonLink
+                              bsStyle='primary'
+                              bsSize='large'
+                              to={`/site/${site._id}/report`}>
+                              <Icon type='print' />View Printable Report
+                            </ButtonLink>
+                          </Col>
+                          <Col xs={7}>
+                            <Icon type='info-circle' /> Click "View Printable Report" to view a summary version of this site analysis suitable for sharing with partners.
+                          </Col>
+                        </Panel>
                       </Col>
-                      <Col xs={6} className='infographic-modeshare-container'>
-                        <div
-                          className='infographic-modeshare-pct-bar'
-                          style={infographicBackground('#3b90c6', regionalTransitModePct)}
-                          />
-                        <p>
-                          {regionalTransitModePct} of all commuters in Washington, DC
-                          take transit to work.
-                        </p>
-                      </Col>
                     </Row>
-
-                    <Row>
-                    <Col xs={6} className='infographic-site-column'>
-                      <h4>Bike Commute</h4>
-                      <div
-                        className='infographic-well'
-                        style={infographicBackground('#f0a800', summaryStats.pctWith30MinBike)}
-                        >
-                        <Icon type='bicycle' />
-                        <span className='number-pct'>{summaryStats.pctWith30MinBike}</span>
-                      </div>
-                      <p>
-                        {summaryStats.pctWith30MinBike} of commuters at
-                        {isMultiSite ? ' these sites ' : ' this site '}
-                        can bike to work in 30 minutes or less.
-                      </p>
-                    </Col>
-                    <Col xs={6} className='infographic-site-column'>
-                      <h4>Carpool/Vanpool Commute</h4>
-                      <div
-                        className='infographic-well'
-                        style={infographicBackground('#ec684f', summaryStats.pctWithRidematch)}
-                        >
-                        <i className='icon-carshare' />
-                        <span className='number-pct'>{summaryStats.pctWithRidematch}</span>
-                      </div>
-                      <p>
-                        {summaryStats.pctWithRidematch} of commuters at
-                        {isMultiSite ? ' these sites ' : ' this site '}
-                        have a rideshare match within 1 mile or less of their homes.
-                      </p>
-                    </Col>
-                    </Row>
-
-                    <Row>
-                    <Col xs={6} className='infographic-modeshare-container'>
-                      <div
-                        className='infographic-modeshare-pct-bar'
-                        style={infographicBackground('#f0a800', regionalOtherModePct)}
-                        />
-                      <p>
-                        {regionalOtherModePct} of all commuters in Washington, DC
-                        take a taxi, motorcycle, bicycle or other means of transportation to work.
-                      </p>
-                    </Col>
-                    <Col xs={6} className='infographic-modeshare-container'>
-                      <div
-                        className='infographic-modeshare-pct-bar'
-                        style={infographicBackground('#ec684f', regionalCarpoolModePct)}
-                        />
-                      <p>
-                        {regionalCarpoolModePct} of all commuters in Washington, DC
-                        carpool to work.
-                      </p>
-                    </Col>
-                    </Row>
+                    <SiteInfographic
+                      commuterCount={commuters.length}
+                      summaryStats={processed.summaryStats}
+                      isMultiSite={isMultiSite}
+                    />
                   </Row>
                 }
               </Tab>
@@ -877,11 +593,11 @@ export default class Site extends Component {
                           <tbody>
                             <tr>
                               <td>% of commuters geocoded:</td>
-                              <td>{pctGeocoded}</td>
+                              <td>{processed.pctGeocoded}</td>
                             </tr>
                             <tr>
                               <td>% of commutes calculated:</td>
-                              <td>{pctStatsCalculated}</td>
+                              <td>{processed.pctStatsCalculated}</td>
                             </tr>
                           </tbody>
                         </Table>
@@ -893,9 +609,9 @@ export default class Site extends Component {
                           <tbody>
                             <tr>
                               <td>% of commuters geocoded:</td>
-                              <td>{pctGeocoded}</td>
+                              <td>{processed.pctGeocoded}</td>
                               <td>% of commutes calculated:</td>
-                              <td>{pctStatsCalculated}</td>
+                              <td>{processed.pctStatsCalculated}</td>
                             </tr>
                           </tbody>
                         </Table>
@@ -996,7 +712,7 @@ export default class Site extends Component {
                   Each row shows how many commuters can commute to work using the currently
                   selected mode up to the travel time listed.
                 </p>
-                <BootstrapTable data={analysisModeStats}>
+                <BootstrapTable data={processed.analysisModeStats}>
                   <TableHeaderColumn dataField='bin' isKey width='150'>Travel Time to<br/>Work (minutes)</TableHeaderColumn>
                   <TableHeaderColumn dataField='cumulative' width='100'>Number of<br/>Commuters</TableHeaderColumn>
                   <TableHeaderColumn
@@ -1011,14 +727,14 @@ export default class Site extends Component {
                 {/***************************
                   Ridematches Tab
                 ***************************/}
-                {!allCommutersGeocoded &&
+                {!processed.allCommutersGeocoded &&
                   <ProgressBar
                     striped
-                    now={pctGeocoded}
+                    now={processed.pctGeocoded}
                     label='Geocoding Commuters'
                     />
                 }
-                {allCommutersGeocoded &&
+                {processed.allCommutersGeocoded &&
                   <div>
                     <FieldGroup
                       label='Map Style'
@@ -1058,7 +774,7 @@ export default class Site extends Component {
                       Each row shows how many commuters have another commuter located within the
                       current distance listed (as the crow flies).
                     </p>
-                    <BootstrapTable data={ridematchingAggregateTable}>
+                    <BootstrapTable data={processed.ridematchingAggregateTable}>
                       <TableHeaderColumn dataField='bin' isKey>Ridematch radius (miles)</TableHeaderColumn>
                       <TableHeaderColumn dataField='cumulative'>Number of Commuters</TableHeaderColumn>
                       <TableHeaderColumn
@@ -1147,14 +863,14 @@ export default class Site extends Component {
                     </Col>
                     <Col xs={12} sm={6}>
                       <h5>Ridematches</h5>
-                      {!ridematches[selectedCommuter._id] &&
+                      {!processed.ridematches[selectedCommuter._id] &&
                         <p>No ridematches within 5 miles of this commuter</p>
                       }
-                      {ridematches[selectedCommuter._id] &&
+                      {processed.ridematches[selectedCommuter._id] &&
                         <div>
-                          <p>{`${ridematches[selectedCommuter._id].matches.length} total matches`}</p>
+                          <p>{`${processed.ridematches[selectedCommuter._id].matches.length} total matches`}</p>
                           <BootstrapTable
-                            data={ridematches[selectedCommuter._id].matches
+                            data={processed.ridematches[selectedCommuter._id].matches
                               .map((match) => {
                                 return {
                                   distance: match.distance,
@@ -1166,7 +882,7 @@ export default class Site extends Component {
                               defaultSortName: 'distance',
                               defaultSortOrder: 'asc'
                             }}
-                            pagination={ridematches[selectedCommuter._id].matches.length > 10}
+                            pagination={processed.ridematches[selectedCommuter._id].matches.length > 10}
                             >
                             <TableHeaderColumn dataField='id' isKey hidden />
                             <TableHeaderColumn dataField='name' dataSort>Matched Commuter</TableHeaderColumn>
@@ -1190,69 +906,31 @@ export default class Site extends Component {
           {/***************************
             Map
           ***************************/}
-          {this.state.mapState !== 'HIDDEN' &&
+          {this.state.mapDisplayMode !== 'HIDDEN' &&
             <Col xs={5}>
-              <div style={this.state.mapState === 'FULLSCREEN' ? {
+              <div style={this.state.mapDisplayMode === 'FULLSCREEN' ? {
                 position: 'fixed',
                 top: 0,
                 left: 0,
                 bottom: 0,
                 right: 0
               } : {height: '600px', marginTop: '1em', marginBottom: '1em'} }>
-                <LeafletMap ref='leafletMap'
-                  style={{ width: '100%', height: '100%' }}
-                  center={position}
-                  bounds={bounds}
-                  zoom={zoom}
-                  >
-                  <TileLayer
-                    url={Browser.retina &&
-                      process.env.LEAFLET_RETINA_URL
-                      ? process.env.LEAFLET_RETINA_URL
-                      : process.env.LEAFLET_TILE_URL}
-                    attribution={process.env.LEAFLET_ATTRIBUTION}
-                    />
-                  {siteMarkers}
-                  {activeTab !== 'ridematches' && commuterMarkers}
-                  {activeTab === 'ridematches' && rideMatchMapStyle === 'normal' &&
-                    commuterMarkers}
-                  {activeTab === 'ridematches' && rideMatchMapStyle === 'marker-clusters' &&
-                    <MarkerCluster
-                      newMarkerData={clusterMarkers}
-                      />
-                  }
-                  {activeTab === 'ridematches' && rideMatchMapStyle === 'heatmap' &&
-                    <Heatmap
-                      intensityExtractor={m => 1000}
-                      latitudeExtractor={m => m.props.position.lat}
-                      longitudeExtractor={m => m.props.position.lng}
-                      points={commuterMarkers}
-                      />
-                  }
-                  {activeTab === 'ridematches' && rideMatchMapStyle === 'commuter-rings' &&
-                    commuterRings}
-                  {activeTab === 'ridematches' && rideMatchMapStyle === 'commuter-rings' &&
-                    commuterMarkers}
-                  {isochrones}
-                  <Legend {...mapLegendProps} />
-                  {this.state.mapState === 'STANDARD' &&
-                    <div style={{ position: 'absolute', right: '10px', top: '10px', zIndex: '10000' }}>
-                      <Button bsSize='small' onClick={() => { this.setState({ mapState: 'HIDDEN' }) }}>
-                        <Icon type='compress' /> Hide Map
-                      </Button>
-                      <Button bsSize='small' onClick={() => { this._setMapState('FULLSCREEN') }}>
-                        <Icon type='expand' /> Fullscreen
-                      </Button>
-                    </div>
-                  }
-                  {this.state.mapState === 'FULLSCREEN' &&
-                    <div style={{ position: 'absolute', right: '10px', top: '10px', zIndex: '10000' }}>
-                      <Button bsSize='small' onClick={() => { this._setMapState('STANDARD') }}>
-                        <Icon type='times' />
-                      </Button>
-                    </div>
-                  }
-                </LeafletMap>
+                <SiteMap ref='map'
+                  commuters={commuters}
+                  isMultiSite={isMultiSite}
+                  polygonStore={polygonStore}
+                  selectedCommuter={selectedCommuter}
+                  site={site}
+                  sites={sites}
+                  activeTab={activeTab}
+                  analysisMode={analysisMode}
+                  analysisMapStyle={analysisMapStyle}
+                  commuterRingRadius={commuterRingRadius}
+                  isochroneCutoff={isochroneCutoff}
+                  rideMatchMapStyle={rideMatchMapStyle}
+                  mapDisplayMode={mapDisplayMode}
+                  setMapDisplayMode={(mode) => { this._setMapDisplayMode(mode) }}
+                />
               </div>
             </Col>
           }
@@ -1279,26 +957,6 @@ CustomHandle.propTypes = {
   value: PropTypes.any
 }
 
-const fillColor = {
-  'blue-incremental': (time) => hslToHex(240, 100, time * 0.00942 + 27.1739),
-  'blue-incremental-15-minute': (time) => hslToHex(240, 100, Math.floor(time / 900) * 8.125 + 30),
-  'blue-solid': '#000099',
-  'green-red-diverging': (time) => hslToHex(time * -0.017391304347826 + 125.217391304348, 100, 50),
-  'inverted': '#000099'
-}
-
-function formatDistance (cell, row) {
-  return humanizeDistance(cell)
-}
-
-function formatPercent (n) {
-  return Math.round(n * 100)
-}
-
-function formatPercentAsStr (n) {
-  return `${formatPercent(n)}%`
-}
-
 function percentBar (n) {
   return <Row>
     <Col xs={9}><ProgressBar now={formatPercent(n)} style={{ marginBottom: 0 }} /></Col>
@@ -1315,130 +973,6 @@ function geocodeConfidenceRenderer (cell, row) {
   } else {
     return 'Not exact'
   }
-}
-
-function getIsochroneLegendHtml ({ analysisMapStyle, isochroneCutoff, analysisMode }) {
-  let html = `<tr><td colspan="2">Travel Time (by ${capitalize(analysisMode.toLowerCase())})</td></tr>`
-  const strategy = getIsochroneStrategies[analysisMapStyle]
-
-  if (strategy === 'single isochrone') {
-    html += `<tr>
-      <td>
-        <div style="border: 1px solid black;">
-          <div style="background-color: ${fillColor[analysisMapStyle]}; opacity: 0.4;">
-            &nbsp;
-          </div>
-        </div>
-      </td>
-      <td>0 - ${shortEnglishHumanizer(isochroneCutoff * 1000)}</td>
-    </tr>`
-  } else if (strategy === 'inverted isochrone') {
-    html += `<tr>
-      <td><div style="border: 3px solid #3388FF;">&nbsp;</div></td>
-      <td>0 - ${shortEnglishHumanizer(isochroneCutoff * 1000)}</td>
-    </tr>`
-  } else {
-    const timeGap = 900
-
-    for (let curTime = timeGap; curTime <= 7200; curTime += timeGap) {
-      html += `<tr>
-        <td style="background-color: ${fillColor[analysisMapStyle](curTime)}; opacity: 0.4;"></td>
-        <td>${shortEnglishHumanizer((curTime - timeGap) * 1000)} - ${shortEnglishHumanizer(curTime * 1000)}</td>
-      </tr>`
-    }
-  }
-  return html
-}
-
-const getIsochroneCache = {}
-
-function getIsochrones ({ analysisMapStyle, analysisMode, isochroneCutoff, polygonStore, site }) {
-  const strategy = getIsochroneStrategies[analysisMapStyle]
-  const cacheQuery = [
-    site.coordinate.lat,
-    site.coordinate.lng,
-    analysisMode,
-    strategy === (strategy.indexOf('minute') > -1) ? strategy : `${strategy}-${isochroneCutoff}`
-  ].join('-')
-  if (getIsochroneCache[cacheQuery]) {
-    return getIsochroneCache[cacheQuery]
-  }
-
-  const allPolygons = Object.values(polygonStore)
-
-  // single extent isochrone
-  if (strategy.indexOf('minute') === -1) {
-    // diff isochrones to get 5 minute isochrones
-    for (let i = 0; i < allPolygons.length; i++) {
-      const curPolygon = allPolygons[i]
-      if (curPolygon.mode === analysisMode &&
-        curPolygon.siteId === site._id &&
-        curPolygon.properties.time === isochroneCutoff) {
-        if (strategy === 'single isochrone') {
-          getIsochroneCache[cacheQuery] = [curPolygon]
-          return [curPolygon]
-        } else if (strategy === 'inverted isochrone') {
-          // diff against massive polygon
-          const hugePolygon = {
-            coordinates: [[
-              [-179.9999, -89.9999],
-              [-179.9999, 89.9999],
-              [179.9999, 89.9999],
-              [179.9999, -89.9999],
-              [-179.9999, -89.9999]
-            ]],
-            type: 'Polygon'
-          }
-          const hugePolygonGeometry = geoJsonReader.read(JSON.stringify(hugePolygon))
-          const invertedGeom = hugePolygonGeometry.difference(reduceAndSimplifyGeometry(curPolygon.geometry))
-          const invertedIsochrone = {
-            geometry: geoJsonWriter.write(invertedGeom),
-            properties: curPolygon.properties,
-            type: 'Feature'
-          }
-          getIsochroneCache[cacheQuery] = [invertedIsochrone]
-          return [invertedIsochrone]
-        }
-      }
-    }
-    // no match found
-    return []
-  }
-
-  const sitePolygons = allPolygons
-    .filter((polygon) => polygon.mode === analysisMode && polygon.siteId === site._id)
-    .sort((a, b) => a.properties.time - b.properties.time)
-
-  let timeGap = 900
-  if (strategy === '5-minute isochrones') {
-    timeGap = 300
-  }
-
-  // diff isochrones to get desired isochrones
-  const isochrones = []
-  let traversedIsochrone
-  sitePolygons
-    .filter((polygon) => polygon.properties.time % timeGap === 0)
-    .forEach((polygon) => {
-      const curFeatureGeometry = reduceAndSimplifyGeometry(polygon.geometry)
-      let isochroneGeometry
-
-      if (!traversedIsochrone) {
-        isochroneGeometry = curFeatureGeometry
-      } else {
-        isochroneGeometry = curFeatureGeometry.difference(traversedIsochrone)
-      }
-
-      isochrones.push({
-        geometry: geoJsonWriter.write(isochroneGeometry),
-        properties: polygon.properties,
-        type: 'Feature'
-      })
-      traversedIsochrone = curFeatureGeometry
-    })
-
-  getIsochroneCache[cacheQuery] = isochrones
-  return isochrones
 }
 
 const getIsochroneStrategies = {
@@ -1468,106 +1002,3 @@ const handleStyle = {
   fontSize: '14px',
   textAlign: 'center'
 }
-
-const homeIconUrl = `${process.env.STATIC_HOST}assets/home-2.png`
-const homeIcon = icon({
-  iconUrl: homeIconUrl,
-  iconSize: [32, 37],
-  iconAnchor: [16, 37]
-})
-
-const homeIconSelectedUrl = `${process.env.STATIC_HOST}assets/home-2-selected.png`
-const homeIconSelected = icon({
-  iconUrl: homeIconSelectedUrl,
-  iconSize: [32, 37],
-  iconAnchor: [16, 37]
-})
-const homeIconSelectedOffset = point(0, -20)
-
-function infographicBackground (color, pct) {
-  return {
-    background: `linear-gradient(to right,
-      ${color} 0%,
-      ${color} ${pct},
-      #b1b3af ${pct},
-      #b1b3af 100%)`
-  }
-}
-
-const isochroneStyleStrategies = {
-  'blue-incremental': {
-    fillOpacity: 0.4,
-    stroke: false,
-    style: (feature) => {
-      return {
-        fillColor: fillColor['blue-incremental'](feature.properties.time)
-      }
-    }
-  },
-  'blue-incremental-15-minute': {
-    color: '#000000',
-    fillOpacity: 0.4,
-    stroke: true,
-    style: (feature) => {
-      return {
-        fillColor: fillColor['blue-incremental'](feature.properties.time)
-      }
-    },
-    weight: 1
-  },
-  'blue-solid': {
-    color: '#000000',
-    fillColor: fillColor['blue-solid'],
-    fillOpacity: 0.4,
-    stroke: true,
-    weight: 1
-  },
-  'green-red-diverging': {
-    fillOpacity: 0.4,
-    stroke: false,
-    style: (feature) => {
-      return {
-        fillColor: fillColor['green-red-diverging'](feature.properties.time)
-      }
-    }
-  }
-}
-
-function onEachFeature (feature, layer) {
-  if (feature.properties) {
-    let pop = '<p>'
-    Object.keys(feature.properties).forEach((name) => {
-      const val = feature.properties[name]
-      pop += name.toUpperCase()
-      pop += ': '
-      pop += name.toUpperCase() === 'TIME' ? humanizeDuration(val * 1000) : val
-      pop += '<br />'
-    })
-    pop += '</p>'
-    layer.bindPopup(pop)
-  }
-}
-
-function reduceAndSimplifyGeometry (inputGeomerty) {
-  const geometry = geoJsonReader.read(JSON.stringify(inputGeomerty))
-  const precisionModel = new geom.PrecisionModel(10000)
-  const precisionReducer = new precision.GeometryPrecisionReducer(precisionModel)
-  return precisionReducer.reduce(simplify.DouglasPeuckerSimplifier.simplify(geometry, 0.001))
-}
-
-const shortEnglishHumanizer = humanizeDuration.humanizer({
-  language: 'shortEn',
-  languages: {
-    shortEn: {
-      y: () => 'y',
-      mo: () => 'mo',
-      w: () => 'w',
-      d: () => 'd',
-      h: () => 'h',
-      m: () => 'm',
-      s: () => 's',
-      ms: () => 'ms'
-    }
-  },
-  spacer: ''
-})
